refactor of code base
