+++ hmm/hmmInput.txt
